+++ src/constants/db.ts
@@ -9,6 +9,7 @@ export enum DbCollection {
 export enum DbSubcollection {
   Profile = 'profile',
   Body = 'body',
+  BodyActivities = 'body_activities',
   Expenses = 'expenses',
   BabyFeeds = 'baby_feeds',
   BabySleep = 'baby_sleep',
@@ -24,7 +25,7 @@ export enum DbDoc {
 
 /** Firestore field names */
 export enum DbField {
-  HeadminickUid = 'headminickUid',
+  AdminUid = 'headminickUid',
   LinkedUid = 'linkedUid',
   UsedAt = 'usedAt',
   Modules = 'modules',